// src/enum/Industry.enum.ts
enum Industry {
  AGRICULTURAL = 'AGRICULTURAL',
  CONSTRUCTION = 'CONSTRUCTION',
  ENERGY = 'ENERGY',
  FINANCE_AND_INSURANCE = 'FINANCE_AND_INSURANCE',
  FOOD = 'FOOD',
  HEALTH = 'HEALTH',
  MANUFACTURING = 'MANUFACTURING',
  MINING_AND_SUBSURFACE = 'MINING_AND_SUBSURFACE',
  PRIVATE_SECURITY = 'PRIVATE_SECURITY',
  WASTE_MANAGEMENT = 'WASTE_MANAGEMENT',
}

export default Industry;