// src/enum/regulatory-election.enum.ts
enum RegulatoryElection {
  AUSTRALIA = 'AUSTRALIA',
  AUSTRIA = 'AUSTRIA',
  BELGIUM = 'BELGIUM',
  CANADA = 'CANADA',
  CHILE = 'CHILE',
  ROATAN_COMMON_LAW_CODE = 'ROATAN_COMMON_LAW_CODE',
  DENMARK = 'DENMARK',
  DUBAI = 'DUBAI',
  ESTONIA = 'ESTONIA',
  FINLAND = 'FINLAND',
  FRANCE = 'FRANCE',
  GERMANY = 'GERMANY',
  HONDURAS = 'HONDURAS',
  HONG_KONG = 'HONG_KONG',
  ICELAND = 'ICELAND',
  IRELAND = 'IRELAND',
  ISRAEL = 'ISRAEL',
  ITALY = 'ITALY',
  JAPAN = 'JAPAN',
  LUXEMBOURG = 'LUXEMBOURG',
  MEXICO = 'MEXICO',
  NETHERLANDS = 'NETHERLANDS',
  NEW_ZEALAND = 'NEW_ZEALAND',
  NORWAY = 'NORWAY',
  PETITION_FOR_TAILORED_REGULATION_GRANTED = 'PETITION_FOR_TAILORED_REGULATION_GRANTED',
  PETITION_FOR_TAILORED_REGULATION_PENDING = 'PETITION_FOR_TAILORED_REGULATION_PENDING',
  POLAND = 'POLAND',
  SINGAPORE = 'SINGAPORE',
  SOUTH_KOREA = 'SOUTH_KOREA',
  SPAIN = 'SPAIN',
  SWEDEN = 'SWEDEN',
  SWITZERLAND = 'SWITZERLAND',
  UK = 'UK',
  USA = 'USA',
}

export default RegulatoryElection;